Move atom feed tests to own file
